add Visualizer component and transfer code
