+++ src/Sequencer/Sequencer.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import * as Tone from "tone";
 import * as d3 from "d3-random";
-import Sketch from "react-p5";
 import Square from "../Square/Square";
+import Visualizer from "../Visualizer/Visualizer";
 import audioProps from "../audioProps";
 
 const Sequencer = () => {
@@ -194,34 +194,9 @@ const Sequencer = () => {
           ))}
         </div>
       ))}
-      <P />
+      <Visualizer />
     </div>
   );
-};
-
-const monoSynthWave = new Tone.Waveform();
-Tone.Destination.connect(monoSynthWave);
-const P = (props) => {
-  const setup = (p5, canvasParentRef) => {
-    // use parent to render the canvas in this ref
-    // (without that p5 will render the canvas outside of your component)
-    p5.createCanvas(400, 100).parent(canvasParentRef);
-  };
-
-  const draw = (p5) => {
-    p5.background(0);
-    p5.stroke(255);
-    let buffer = monoSynthWave.getValue(0);
-    for (let i = 0; i < buffer.length; i++) {
-      let x = p5.map(i, 0, buffer.length, 0, p5.width);
-      let y = p5.map(buffer[i], -1, 1, 0, p5.height);
-      p5.point(x, y);
-    }
-    // NOTE: Do not use setState in the draw function or in functions that are executed
-    // in the draw function...
-    // please use normal variables or class properties for these purposes
-  };
-  return <Sketch setup={setup} draw={draw} />;
 };
 
 export default Sequencer;
